requirements.txt
agno
google-genai
requests
python-dotenv
rich
